Raise the actual record when innodb transaction times out on a row lock
Summary:
When a transaction times out on waiting for a row lock in innodb, we will print
out the primary key of the row. This is implemented similarly to populating the
LOCK_DATA in the I_S table information_schema.innodb_locks.

Privacy: we should be fine to print out the record that the transaction is
waiting on, as we should have the privilege to look at the record already.

Example:
  ERROR HY000: Lock wait timeout exceeded; try restarting transaction: Timeout on record (3) in index: test/t1.a

Note: this is implemented in the Innodb SE. RocksDB SE will be a separate diff.

Test Plan:
Plenty of existing tests on the lock_wait_timeout. Also added a test case for
very large key.

Reviewers: flupps, abhinavsharma, herman

Reviewed By: herman

Subscribers: phabricatorlinter, jkedgar, webscalesql-eng@fb.com

Differential Revision: https://phabricator.intern.facebook.com/D5606554

Tasks: T17664677

Signature: t1:5606554:1503527276:0d8f5bcc9b94692e903eef7cabe941c847075ca5