Adding high_priority syntax in DDL commands
Summary:
This diff introduces `HIGH_PRIORITY` syntax in some DDL commands and optimize
command. The purpose is the same as the system variable `high_priority_ddl`
which is described in D4784076.

The list of commands that support `HIGH_PRIORITY` syntax is:
- create/drop/alter/truncate table
- create/drop index
- create/drop trigger
- optimize table

With this diff, there are two ways of issuing high priority DDLs:
  e.g. `CREATE HIGH_PRIORITY INDEX ...`

Recap of the high_priority ddl feature:

- Only the shared locks (read/write) will be affected, which are DML queries
  such as select, insert, update, etc. Any lock higher than upgradable MDL
  (i.e. from conflicting DDL) will not be affected.
- When killing the blocking connections, their transactions will automatically
  be rolled back. This is the same behavior as `kill processlist_id` command.
- The DDL will kill blocking connections at the end of the timeout instead of
  instantly. Admin can control the waiting period by setting the session
  lock_wait_timeout before issuing the DDL.

Reviewed By: abhinav04sharma

Differential Revision: D5635549

fbshipit-source-id: e51ed02