Fix flakiness in rocksdb.ttl_secondary
Summary:
Some tests in rocksdb.ttl_secondary were flaky from the use of UNIX_TIMESTAMP for testing explicit timestamps. During an update, if the secondary key does not change (because the UNIX_TIMESTAMP returned the same value), then we never enter the engine to update the TTL values. Fix this by adding 1 to UNIX_TIMESTAMP when doing updates.

This same fix was already present for another test in the file, but the fix wasn't applied to the whole file.

Squash with: D5463064

Reviewed By: hermanlee

Differential Revision: D5693233

fbshipit-source-id: c107fdc