Disable checking tables with #sql during startup
Summary:
When myrocks starts up, it defaults to verifying all tables in the myrocks data dictionary has matching .frm files in mysql. However, there are cases where temporary tables are created with #sql in the filename. For these tables, the check against a .frm file should be removed. We currently ignore tables that start with #sql, but some distributions like MariaDB may append a prefix to the tablename, so change it to search for #sql in any part of the name.

ttps://github.com/facebook/mysql-5.6/issues/193

Reviewed By: hermanlee

Differential Revision: D5736715

fbshipit-source-id: b6e8096