Exit if mysqlbinlog process exits prematurely
Summary:
For some reason, mysqlbinlog does not always start up and connect to the server properly, and we end up with just 9 processes connected to the server, which leads to the test timing out.

In a previous attempt https://github.com/facebook/mysql-5.6/commit/7815a4f4fd0e274fbe6fddf11dd9273a9bf3e6dc, we tried to collect the exit code of the mysqlbinlog process so that we knew what was going on, but it was still hung because we never exited when we detected the process was gone. Add in an exit.

Squash with: D5419248

Test Plan: mtr

Reviewers: herman

Reviewed By: herman

Subscribers: webscalesql-eng@fb.com

Differential Revision: https://phabricator.intern.facebook.com/D5738976

Tasks: T21123122

Signature: 5738976:1504124463:b7a7e88a9da3496eafd756f3ed47135b16887a6e