Initialize SstFileManager with proper delete rate
Summary:
When we Create SstFileManager it checks if there are existing files in trash and delete them using the assigned delete rate.
Since we create the object then assign the rate later we end up deleting all files immediately.

The fix is to create the SstFileManager object with the assigned delete rate
Closes https://github.com/facebook/mysql-5.6/pull/712

Reviewed By: yoshinorim

Differential Revision: D5879960

Pulled By: IslamAbdelRahman

fbshipit-source-id: 8332c5e