Add security to the rpc_role query attribute
Summary: The rpc_role query attribute allows a user (i.e. the MyRouter user) to tell the MySQL server to run the query as someone else.  This gives great power to users and we need to have this protected.  This change hooks this system in to the GRANT PROXY privilege.  If a user has the right to proxy as another user then they can use the rpc_role query attribute to run a query as that user.  Note that 'root' defaults to having proxy privileges for all other users.

Reviewed By: anca-agape

Differential Revision: D6009461

fbshipit-source-id: ebc02bd