Fix rocksdb.bulk_load_errors test flakiness
Summary:
The current test checks the log on disconnect to ensure an error message
was printed out for bulk load disconnect. However, the test can
sometimes fail to find the error message if the log is processed before
the error message is printed.

Add a wait to the test for the thread count to drop back down to the
original.

Reviewed By: asandryh

Differential Revision: D6027184

fbshipit-source-id: 3195ad7aff7