Option to shutdown master if semi-sync is being switched off while there are active un-acked trxns
Summary:
Semi-sync master can be switched off manually or because of a timeout while
waiting for an ack.

This change introduces a new variable rpl_semi_sync_master_crash_if_active_trxs
which when set to true will cause the master to shutdown (after printing an
error msg) if semi-sync is being switched off while there are active un-acked
transactions. This prevents any un-acked trx from committing on the
master.

Reviewed By: hermanlee

Differential Revision: D5841339

fbshipit-source-id: c378441