adding numa_nodebind=${node} option to mysqld_safe
Summary:
Automation will pass in the right value to this option if
it needs to bind mysqld to a particular socket. That will happen
via systemd scripts.

This is only adding the capability on the mysqld_safe side to be able to
pass in numactl to the server. The turning on of this feature will need more
testing and slow roll, and that will happen in different diffs on the ops repo.

Differential Revision: D6067155

fbshipit-source-id: a4242d2