Add a hashed, hierarchical, wheel timer implementation
Summary:
In order to implement idle timeouts on detached sessions we need something inside MySQL that is lightweight and can handle calling events in the future with very little cost for cancelling or resetting the event.  A hashed, hierarchical, wheel timer fits the bill.  This implementation was derived from the Folly implementation (https://github.com/facebook/folly/blob/master/folly/io/async/HHWheelTimer.h) but doesn't depend on event bases - instead it hooks into the exisiting my_timer implementation in MySQL.

By default the timers are grouped into 10ms buckets (the 'hashed' part), though the size of the buckets is configurable at the creation of the timer.  Each wheel (the 'wheel' part) maintains 256 buckets and cascades to the wheel below it (the 'hierachical' part).  The lowest wheel represents the next 2560 milliseconds (256 buckets at 10 ms each).  The buckets in the second wheel represents 2560 ms each, the buckets in the third weel represent 655360 ms each, etc.)

Reviewed By: djwatson

Differential Revision: D6199806

fbshipit-source-id: 5e1590f