Put RPC_ID into the session state portion of the OK packet
Summary: For future expansion we want to move the RPC_ID from the 'info' section to the 'session state' section of the OK packet.  This diff adds it to the 'session state' section without removing it from the 'info' section.  Once this has landed we can update MyRouter to pull it out of the 'session state' section and once that is landed we can remove the code that puts it in the 'info' section.

Test Plan: MTR.  I know that main.com_rpc reads the rpc_id from the 'session state' section because during debugging the test was constantly failing.

Reviewers: #dbclient, herman, mung, anca

Reviewed By: anca

Subscribers: webscalesql-eng@fb.com

Differential Revision: https://phabricator.intern.facebook.com/D6096092

Tasks: T20014998

Signature: 6096092:1508801267:2bf2e0cb79bce9d6b968968c8014a702a579a3a1