Put RPC_ID into the session state portion of the OK packet
Summary: For future expansion we want to move the RPC_ID from the 'info' section to the 'session state' section of the OK packet.  This diff adds it to the 'session state' section without removing it from the 'info' section.  Once this has landed we can update MyRouter to pull it out of the 'session state' section and once that is landed we can remove the code that puts it in the 'info' section.

Reviewed By: anca-agape

Differential Revision: D6096092

fbshipit-source-id: 2993f3a