Fix crash during shutdown in audit plugin code path
Summary:
On server shutdown, there exists a race condition that can trigger a segfault if an audit plugin is installed. The root cause because the `THD::audit_class_plugins` array is accessed concurrently without locks.

This is the sequence of actions that can trigger a crash:
1. The shutdown thread calls close_connections on all connected clients.
2. The connection thread will call into the audit plugin because it will receive a network error from `close_connections`, and will log the error into the audit plugin. The thread then adds the plugin to `THD::audit_class_plugins`, but it has not updated `audit_class_mask` yet.
3. The shutdown thread sees that `audit_class_mask` is not set, and also adds the plugin into `THD::audit_class_plugins`, but it has only extended the array, without setting the value of the 2nd element of the array.
4. The connection thread then goes ahead and updates `audit_class_mask` and finishes `acquire_plugins`. It then calls `mysql_audit_notify` and sees the array of size 2, and when it loops over the 2nd element, it segfaults because the value is uninitalized.

The fix is to add a lock per THD that protects these data structures. They should be relatively uncontended since the shutdown thread is the only thread that calls `acquire_plugins` on another THD.

The corresponding code in 5.7/8.0 has been largely refactored, but I suspect this will still be an issue since it doesn't seem like the corresponding data structures are synchronized.

Test Plan:
I was able to reproduce the issue consistently with this patch: P58457941. I've put sleeps to trigger the sequence of events explicitly. This did not repro after adding the lock.

This should fix these crashes we see in production: https://fburl.com/3ib2j8ou

Reviewers: herman, aliaksei

Reviewed By: aliaksei

Subscribers: aliaksei, webscalesql-eng@fb.com

Differential Revision: https://phabricator.intern.facebook.com/D6103777

Signature: 6103777:1508521075:39b55763e05f10ec8c2569ffbf1604ce306350a2