Skip scan does not work with bloom filters in MyRocks
Summary:
During skip scan, we don't reset the end range when we iterate to the next prefix to scan. For MyRocks, end range is used to determine whether bloomfilters can be used or not, so because it was not reset, MyRocks was making the wrong decisions about whether bloomfilters are usable or not.

The fix is to reset it before every prefix scan.

Closes https://github.com/facebook/mysql-5.6/issues/744

Test Plan: I've added an opt file to rocksdb.optimizer_loose_index_scans to enable bloomfilters

Reviewers: herman, yoshinori

Reviewed By: yoshinori

Subscribers: webscalesql-eng@fb.com

Differential Revision: https://phabricator.intern.facebook.com/D6140923

Signature: 6140923:1508898895:2da2387e4454f9d89c8569168b8ac2836b012db8