Implement inplace alter for auto_increment (#738) (#738)
Summary:
This implements inplace alter for changing auto_increment values. It works by reading the last value of the auto increment index and compares that with the request auto_increment value in the alter statement. The larger of the two values is set as the new autoincrement value for the table.

To handle partitions, every partition is handled as a separate table. Since we call `prepare_inplace_alter_table` for every partition but `commit_inplace_alter_table` only once, we load the last value in the index during prepare and write into the data dictionary on commit.
Closes https://github.com/facebook/mysql-5.6/pull/738

Differential Revision: D6083120 (https://github.com/facebook/mysql-5.6/commit/9f53fae746084823d35177b70a314d14c93a441b)

fbshipit-source-id: 98bd5758fe2