Fixed rocksdb_bulk_load_allow_unsorted option callback.
The option uses a 'set' callback to set/return an error to the client which is
an invalid thing to do. This is proven by the debug assertion that is hit on a
debug build when executing the sequence in the rocksdb.bulk_load_unsorted_errors
test. The proper way to return an error from a SET is to use a 'check' callback
instead. This patch changes the 'set' to a 'check' and removes the no-debug
constraint on the bulk_load_unsorted_errors test.