Fix test flakiness in singledelete_idempotent_recovery by enforcing master_pos_wait()
Summary:
Add an extra transaction on the master to force the slave to comply with
master_pos_wait(). Since the test crashes the slave after it has
processed the binlog entry and then truncates the binlog before
recovery, the slave's master position can be in flux when it's being
checked via master_pos_wait(). This can cause the select
master_pos_wait() function called during --sync_slave_with_master in
mysqltest.cc to return early.

When this happens, the test could delete the row before the binlog is
replayed, causing the row to reappear after the delete.

The extra transaction forces an update to the master position, which the
slave has never seen, so master_pos_wait() will wait until the new
transaction shows up before returning.

Reviewed By: lth

Differential Revision: D6266910

fbshipit-source-id: 781bf82eded