Add mysqldump option to print ordering key
Summary:
The --print-ordering-key option to mysqldump causes mysqldump to get the
key which is used to order the rows for each table, and print it to the
dump file.

This will help to efficiently run a diff on two mysqldumps of the same
database at different points of time.

Reference patch: https://github.com/facebook/mysql-5.6/commit/0f2578a

Differential Revision: D6364804 (https://github.com/facebook/mysql-5.6/commit/42333e7696d2f6cb4710fab49dfa6416b5b288fd)

fbshipit-source-id: f69a97d906e