Fix setting index_info flags correctly for version INDEX_INFO_VERSION_TTL
Summary:
When tables are setup with TTL and created with index info version
INDEX_INFO_VERSION_TTL, and then MyRocks is upgraded to a build that
uses INDEX_INFO_VERSION_FIELD_FLAGS, the index_field is not set
correctly for tables that have INDEX_INFO_VERSION_TTL.

Fix this by setting the index_flags to indicate TTL is used
when ttl_duration is detected to be greater than 0.

Test Plan:
Create a database with the old format, then upgrade to the new. No mtr
test will be provided given the complexity of the upgrade.

Reviewers: aliaksei, mung, yoshinori, rtm

Reviewed By: aliaksei

Subscribers: webscalesql-eng@fb.com

Differential Revision: https://phabricator.intern.facebook.com/D6383831

Tasks: T23748488

Signature: 6383831:1511283933:7cc4f0088477aa8754181a858aff932ba6d79ea2