Using table names instead of table IDs to identify dependencies
Summary:
Table IDs are not guaranteed to be unique. They are assigned to tables
using an in-memory counter that is incremented everytime a new table is opened.
So, under certain circumstances like master reboot the same table can have
multiple table IDs.

This change uses database name + table name to identify dependencies instead of
table IDs.

Reviewed By: anirbanr-fb

Differential Revision: D6389969

fbshipit-source-id: f90c925