Make mysqldump not lock all tables with --single-transaction
Summary:
Change the behavior of --single-transaction to use

	START TRANSACTION WITH CONSISTENT INNODB SNAPSHOT

to begin transactions, which returns the binlog file and offset. This
is later used if --master-data is specified and becomes part of the
output (otherwise it just avoids the lock tables).

Reference patch: https://github.com/facebook/mysql-5.6/commit/32b49db

Depends on D6437300

Reviewed By: Tema

Differential Revision: D6437452

fbshipit-source-id: 39671acbad6