Change the response attribute packet structure so it can easily be skipped
Summary: The response attributes implementation did not follow the standard correctly.  Instead of a length encoded integer representing the size of the rest of the data after the type byte it used a length encoded integer of the number of entries.  This would make it tricky for a client to skip, especially if it doesn't know the format.  Change this by adding a length encoded size.

Test Plan: MTR

Reviewers: jprahman, herman, anca, mung

Reviewed By: jprahman

Subscribers: webscalesql-eng@fb.com

Differential Revision: https://phabricator.intern.facebook.com/D6431642

Signature: 6431642:1511915726:a5dd17d13fc58dc25ed83f0502ac77183669a71c