Change the response attribute packet structure so it can easily be skipped
Summary: The response attributes implementation did not follow the standard correctly.  Instead of a length encoded integer representing the size of the rest of the data after the type byte it used a length encoded integer of the number of entries.  This would make it tricky for a client to skip, especially if it doesn't know the format.  Change this by adding a length encoded size.

Reviewed By: jrahman

Differential Revision: D6431642

fbshipit-source-id: 87c8171