Basic Async MySQL Client Support
Summary:
This diff builds loosely upon the 5.1 client work at Facebook, and
brings async to 8.0.  It is not fully feature complete (no SSL. no
compression, only supports the default native auth) but it is a good
first step, is sufficient for our needs, and we can extend it.

This also adds a #define to indicate the nonblocking client is available.

Reference patch: https://github.com/facebook/mysql-5.6/commit/3c4e084

Reviewed By: abal147

Differential Revision: D6476178

fbshipit-source-id: 438eeda7354