Revert "Raise the actual record when innodb transaction times out on a row lock"
Summary:
The original change integrated the information_schema show lock code
into the lock timeout codepath to try and determine the row record being
locked. However, there seems to be a problem where the record being
retrieved can be NULL, causing MySQL to crash when deciphering the
record. This diff reverts the original change and removes attempting to
find the row lock in the timeout path.

Differential Revision: D6465891

fbshipit-source-id: 32d3169