Implement Async SSL Handshake
Summary:
This is a port of https://github.com/facebook/mysql-5.6/commit/5d914eb. The SSL establishment codepaths has been greatly refactored, so most of the summary in the original diff do not apply.

This adds a cli_establish_ssl_nonblocking counterpart that performs the SSL handshake in a reentrant manner. This is done by storing the SSL object and state on the mysql connect context struct. The original diff put it on the NET struct, which I think is the wrong place for it.

The ssl_handshake_loop function was modified to return VIO_SOCKET_WANT_WRITE or VIO_SOCKET_WANT_READ if nonblocking mode is requested.

Reference patch:
https://github.com/facebook/mysql-5.6/commit/5d914eb
https://github.com/facebook/mysql-5.6/commit/fecdbfb
https://github.com/facebook/mysql-5.6/commit/bee9c39

Reviewed By: abal147

Differential Revision: D6524530

fbshipit-source-id: 050b3b84cf8