Add facility to re-use SSL_CTX across connections
Summary:
Every connection, MySQL recreates the SSL_CTX object.  This is
bad and unnecessary.  This diff adds the ability to re-use an SSL
context object between connections.  This is roughly a 50% increase in
throughput making SSL connections.  In combination with re-using SSL
sessions, the net speedup is roughly 7-8x.

This facility can be used in one of two ways; one is to "take ownership"
of the SSL_CTX in an existing MySQL connection (simple) or to create a
context by hand, therefore bypassing many of the MySQL "options" fields
relating to verification, certs, etc.  This is not ideal but avoids a
larger refactoring and public API change to support some kind of
"initialize an OpenSSL context with these options" which would be
complicated and potentially difficult to reconcile with yassl.

Reference patch:
https://github.com/facebook/mysql-5.6/commit/4c288ca

Differential Revision: D6536123

fbshipit-source-id: a8018bda625