Fix the key portion of the response attribute
Summary:
The original code used a constexpr with STRING_WITH_LEN() - which used sizeof(x)-1 instead of strlen(x)-1 to get the size of the input buffer.  Since the input was a constexpr auto, it got 'const char *' instead of 'char[]', so the size always came back as 8 (-1) which meant that only the first seven bytes of the key were sent.  Change the code to follow the pattern in MySQL where the hardcoded string is supplied directly to STRING_WITH_LEN.

Squash with: D6353096

Reviewed By: abal147

Differential Revision: D6572523

fbshipit-source-id: 57061d9