Fix Async Compression read bug
Summary:
net_read_packet_nonblocking does not set return values when
returning ASYNC_NOT_READY. This needs to respected by all similar
functions.

my_net_read_compressed_nonblocking calls net_read_packet_nonblocking
repeatedly in the case of big (>16mb) packets. In this case if the
first read succeeds, but the remaining data is not available in the
socket, the function returns with ASYNC_NOT_READY.

The calling function in cli_safe_read_nonblocking unconditionally
adds the bytes read to a variable, so in this case if the function
exits early after completing a read the packet size is double counted.

In pretty much every case, this will result in the returned packet
size being some multiple of 16mb greater than what is intended, most
likely resulting in seg faults.

Fix is to not set the return values when not ready, but also not apply them
in callsite

Reviewed By: lth

Differential Revision: D6608202

fbshipit-source-id: 286bb43