Add MYSQL_COMPRESSION_NONE to compress enum
Summary:
Add mysql_compression_none option for setting
compression type. Makes other APIs built off this a little
cleaner as they can unconditionally set some value.

This is safe because the wire protocol that dictates the
values chosen are the names of the compression libs, this
is a client side only change.

Reviewed By: lth

Differential Revision: D6673749

fbshipit-source-id: 16aa400