Don't allow rpc_id to be used from a different host
Summary: When an detached session is created (i.e. an rpc_id is returned), don't allow it to be used from any connection except from the same machine.  Ideally we would only allow it from the same process, but we don't have the infrastructure for that, so restricting it to the same machine is sufficient for now.  To do this we grab the hostname or ip of the connection when the detached session is created and then always compare this to the current session's hostname or ip when executing a command when the rpc_id query attribute is present.

Test Plan: New test case in main.com_rpc

Reviewers: bert, jprahman, herman, aasogamo, anca, mung, abalsara

Reviewed By: anca

Subscribers: webscalesql-eng@fb.com

Differential Revision: https://phabricator.intern.facebook.com/D6739690

Tasks: T20014660

Signature: 6739690:1516219088:fb62cfc6989cdf014e490a92504f8cc0e4a1cf04