Fix high_priority_ddl from timing out
Summary:
When high_priority_ddl is used, it is possible for the lock acquisition
to timeout. Before the connections are killed, m_wait.timed_wait() is
called with set_status_on_timeout. Once the status is set, even if the
lock frees up, it won't be granted to the pending mdl_ticket. The
existing code resets m_wait status before killing the other connections,
but if the lock is granted before m_wait is reset, then the grant fails
and the ddl still times out.

The solution is to prevent m_wait from being set to timeout before the
kill connection occurs. Then it never needs to be reset.

Reviewed By: yashtc

Differential Revision: D6744017

fbshipit-source-id: f598ab4