Safely skip over unsupported session tracking types
Summary:
When we backported session tracking from 5.7, we did not backport all types, which is fine. However, we weren't skipping over them properly, since we just break out prematurely.

The fix is to move the cases down with the default cases where they will safely get skipped. Note that the assert won't fire because we still have all the enums defined properly.

Closes https://github.com/facebook/mysql-5.6/pull/781
Closes https://github.com/facebook/mysql-5.6/issues/766

Reviewed By: jkedgar

Differential Revision: D6874150

fbshipit-source-id: 21532fe