Add rocksdb transaction options for write policy and commit time writebatch
Summary:
This adds two sysvars for MyRocks that controls rocksdb transactions:

TransactionDBOptions::write_policy - This is a read only variable that determines when data is written into the database. The default is COMMITTED which means data is written at commit time.
TransactionOptions::use_only_the_last_commit_time_batch_for_recovery - This is a dynamic variable that determines whether the commit time write batch is written into the database or not. If the commit time write batch is only useful for recovery, writting to WAL is enough.

Reviewed By: hermanlee

Differential Revision: D6905729

fbshipit-source-id: 410bd83