Change the format of query response attributes
Summary:
The original implementation of query response attributes attempted to group them all into one entry in the OK packet.  Later it was discovered that SESSION_TRACK_SYSTEM_VARIABLES, which has a similar type of data, puts each variable in its own entry in the OK packet.  Adjust the format of the query response attributes to match the system variables.

Note this is expected to be safe because we do not believe that any clients are currently using the query response attributes.

Test Plan: MTR

Reviewers: mung, abalsara

Reviewed By: mung

Subscribers: webscalesql-eng@fb.com

Differential Revision: https://phabricator.intern.facebook.com/D6993892

Signature: 6993892:1518730368:c110d975235605b4dfb25cf91c0e70abd2b4750d