Add ADMIN PORT privilege to allow non-super users to use admin port
Summary:
Currently to access the admin port the user must be a super
user, which has other implications. There are situations where we want
to allow a normal user to access MySQL via the admin port, without
resorting to using the admin user. Here, we add a new privilege, that
can be granted to a user and then allow that user to use the admin port.

Reviewed By: Tema

Differential Revision: D7037574

fbshipit-source-id: 6210de4