Update rocksdb.check_ignore_unknown_options test
Summary:
Due to https://github.com/facebook/rocksdb/pull/3527, ignore unknown options only takes effect if the rocksdb options file is from a higher version. Update the test to change the version of the file to something large.

I've also updated the sort because it wasn't actually sorting the file by its number. The restoration of the file was also doing nothing because a new option file gets created with the invalid option removed already.

@update-submodule: rocksdb

Differential Revision: D7071639