Fix main.ctype_uca test
Summary:
The initial diff to handle idle timeout caused a failure in main.ctype_uca to show up.  The problem wasn't actually in ctype_uca, but that test turned out to be the first test that was run after com_rpc that shut down the server.  Shutting down the server after com_rpc showed that we were leaving around some sessions that were keeping a reference count on the MyISAM plugin.

This change fixes the problem of sessions being left around.  The main issue is that we were enabling idle timeouts on some sessions that we shouldn't have.  We only want idle timeouts to occur on sessions that are not already connected to some THD.  We now have an accurate way of determining that.

Also we were not cancelling all idle timeouts during shutdown which caused some detached sessions to stay active during shutdown, keeping references to plugins around.

In addition a call to shutdown the mysqld server was added to the tests so that future problems related to this will point to the correct test.

Squash with D6989712

Reviewed By: Tema

Differential Revision: D7173836

fbshipit-source-id: 5f1a152