SQL to create and manage explicit snapshots
Summary:
This change introduces the following SQL statements:

- START TRANSACTION WITH SHARED <engine> SNAPSHOT - This works just like
transactions with consistent snapshot but the snapshot can be reused by other
transactions (see next SQL statement). The binlog pos, gtid executed and ID of
the snapshot is printed in the output.

- START TRANSACTION WITH EXISITING <engine> SNAPSHOT X - This works by using an
existing snapshot (created using the previous statement) instead of creating a
new one. X is the snapshot ID.

- CREATE EXPLICIT <engine> SNAPSHOT - This creates a snapshot of the current
state of the DB and associates it with the connection. The snapshot can be
reused by other connections. The output contains the binlog pos, gtid executed
and a snapshot ID. It works just like transaction with a shared snapshot but it
operates on the connection level and does not begin a transaction. After this
statment, all transactions in the connection will use the created snapshot until
it is released.

- ATTACH EXPLICIT <engine> SNAPSHOT X - This is analogous to transaction with
existing snapshot but operates on the connection level. X is the snapshot ID.

- RELEASE EXPLICIT <engine> SNAPSHOT - This releases the snapshot associated with
the connection. After this statment, all transactions in the connection will
create their own snapshots.

Reviewed By: hermanlee

Differential Revision: D6818855

fbshipit-source-id: 0b671a3