Prevent cert/key mismatch is combined file replaced between cert load and key load
Summary:
MySQL was using the functions that take a file as input for setting the
ssl cert and key.
There is a race condition if the cert file gets replaced between the
moment the cert is loaded and the key is loaded.
Fix: open file once and load cert and key. Used BIO functions that work
on file descriptors.

Reference patch: https://github.com/facebook/mysql-5.6/commit/53d0673

Reviewed By: abal147

Differential Revision: D7299948

fbshipit-source-id: 96022f317ca