Return state information for detached sessions
Summary:
The original implementation to return session information did not work for detached sessions.  Modify the detached sessions code to put the session state change information in the right location so that it gets returned correctly.

When using detached sessions there are two THDs - the one associated with the connection thread and the one associated with the detached session.  The prior diffs tried to handle this by having a session_tracker_ member variable that pointed into the connection thread's THD to update the session_tracker information correctly.  This was not comprehensive enough as some session tracking changes occurred at a much deeper level.  Modify the session tracking so that it always updates the copy in the detached session so that it will correctly get aggregated and returned to the client.

Once this was working I discovered that occasionally sessions that were detached because of a state change go reattached to the connection.  This is incorrect as there is no way to determine if the state has reverted back to before the state change.  I made additional changes to make sure that a session that has had state change never becomes the default session for a connection.  And once this change was made the existing tests had to be updated because some sessions (that originally had incorrectly become reattached) now showed up in the output results and needed to be specifically killed.

Reviewed By: anca-agape

Differential Revision: D7504732

fbshipit-source-id: e0de551