Removing empty directories for dropped DBs
Summary:
The issue with DROP DATABASE is that it is not redo logged. InnoDB
tablespaces are removed during the apply-log phase, but empty
database directory is left.

Since DDL is not possible during the copying of non-InnoDB files, we
can rely on the presence of db.opt file backup is prepared.

The fix is to remove empty database directories when performing
prepare step.

Ported from the fix for xtrabackup bug1461735. Also ported fix for bug1123335 to satisfy the new testcase.

https://bugs.launchpad.net/percona-xtrabackup/+bug/1461735
https://bugs.launchpad.net/percona-xtrabackup/+bug/1123335

Reviewed By: hermanlee

Differential Revision: D7419025

fbshipit-source-id: d1cf383