Fix Valgrind errors on main.com_rpc* tests
Summary: [COM_RPC] When profiling is turned on and a COM_QUERY_ATTRS command comes in, the original THD's profiling object was initialized, but the COM_RPC's THD was used later on, meaning we were using unitialized data.  Added code to initialize the profiling object of the COM_RPC thread.

Reviewed By: lth

Differential Revision: D7489322

fbshipit-source-id: 6e00dd0