Response attribute handling in MTR allowed a partial match when it shouldn't have
Summary: The code to read response attributes for MTR testing used a strncmp() call using the length of the attribute.  This means that when 'wait_ti' was being returned incorrectly it still was considered a match for 'wait_timeout'.  The 'wait_ti' bug was fixed awhile ago, but the test framework code was never fixed.

Reviewed By: jrahman

Differential Revision: D7501028

fbshipit-source-id: 523c57b