Add variable to force range plans if force index is used
Summary:
Currently, when force index is used, full table scans are highly penalized, but it is still possible to do a full index scan which is expensive. In many cases, we only use force index when we know there is better way to traverse the index, and not necessarily because we want a full index scan.

To address these cases, add a session variable called optimizer_force_index_for_range, which when turned on, will try to search for a 'range' plan before falling back to an index plan.

Test Plan: added new test

Reviewers: herman

Reviewed By: herman

Subscribers: ebergen, dmituzas, webscalesql-eng@fb.com

Differential Revision: https://phabricator.intern.facebook.com/D7526748

Tasks: T12385775

Signature: 7526748:1523298651:3be86ca66dcebe805a9df356ed7923e285ee5c14