Disable bloomfilter when seeking to index+1
Summary:

https://github.com/facebook/mysql-5.6/issues/809

When seeking to a infimum or supremum records of an index, the record's
index value might be different from the key definition. If the
bloomfilter is also set to 4 bytes cap, bloomfilters would be enabled.

i.e. Key definition index value is 260, but we seek to the supremum
record with index value 261. Since bloomfilter ends up being enabled,
Seek(261) would use 261 as the matching key and none of the keys in the
table starting with index value 260 would be found. The result is
selecting those rows fail.