Add variable to disable full table/index scans
Summary:
There are cases where it is always unacceptable for a client to be doing full table scans. To prevent this from happening, add a new variable optimizer_full_scan. When it is turned off, it will return a new error code ER_FULL_SCAN_DISABLED.

For best results, use with optimizer_force_index_for_range. Currently, it does not fallback to alternative plans because we check at the end of query planning.

Reviewed By: hermanlee

Differential Revision: D7528820

fbshipit-source-id: 94fcd07