Delete pid file when crashing after timeout with unacked trx
Summary:
If rpl_semi_sync_master_crash_if_active_trxs = true, the server crashes after a
timeout if there are any active unacked transactions. mysqld_safe expects the
pid file to be deleted if the expectation is to not bring the intance back up
again. This change deletes the pid file before crashing on this condition.

Squash with D5841339

Reviewed By: Tema

Differential Revision: D7747659

fbshipit-source-id: a94df24